requirements.txt
python-dotenv
pyrogram
tgcrypto
requests
speedtest
heroku3
aiohttp
pymongo